requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
constantly==15.1.0
coverage==6.3.1
cryptography==36.0.1
daphne==3.0.2
Django==2.2
django-countries==7.2.1
django-crispy-bulma==0.2
django-crispy-forms==1.7.2
django-model-utils==4.2.0
django-notifications-hq==1.6.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jsonfield==3.1.0
notifications==0.3.2
Pillow==9.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
zope.interface==5.4.0
